fixed minor bugs

+++ Firmware/src/utils.cpp
@@ -37,8 +37,8 @@ esfericalAngles_t getSavedOffsets(void)
 {
     esfericalAngles_t offsets;
     config.begin("config", false);
-    offsets.azimuth = config.getDouble("offset_az", 0.0);
-    offsets.elevation = config.getDouble("offset_el", 0.0);
+    offsets.azimuth = config.getFloat("offset_az", 0.0);
+    offsets.elevation = config.getFloat("offset_el", 0.0);
     config.end();
     ESP_LOGI(TAG, "Updated Offsets - Az: %.3f [°], El: %.3f [°]", offsets.azimuth, offsets.elevation);
     return offsets;
